notification-service/requirements.txt
Django==4.2
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
python-decouple==3.8
celery==5.3.1
twilio==8.3.0